refactor: 모든 인수테스트의 중복코드를 AcceptanceTest.java로 이동
