+++ src/test/java/com/saebyeok/saebyeok/acceptance/CommentAcceptanceTest.java
@@ -3,19 +3,11 @@ package com.saebyeok.saebyeok.acceptance;
 import com.saebyeok.saebyeok.domain.Comment;
 import com.saebyeok.saebyeok.dto.ArticleResponse;
 import com.saebyeok.saebyeok.dto.ExceptionResponse;
-import com.saebyeok.saebyeok.infra.JwtTokenProvider;
-import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.security.test.context.support.WithUserDetails;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.jdbc.Sql;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -23,35 +15,21 @@ import java.util.List;
 import java.util.Map;
 
 import static com.saebyeok.saebyeok.domain.CommentTest.*;
-import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 
-@WithUserDetails("123456789")
-@ActiveProfiles("test")
-@Sql({"/truncate.sql", "/emotion.sql"})
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class CommentAcceptanceTest {
-    private static final String API = "/api";
+class CommentAcceptanceTest extends AcceptanceTest {
     private static final Long ARTICLE_ID = 1L;
     private static final Long EMOTION_ID = 1L;
     private static final List<Long> SUB_EMOTION_IDS = Arrays.asList(1L, 2L);
     private static final Long MEMBER_ID = 1L;
     // TODO: 2020/07/20 MEMBER_ID는 data.sql에 있는 맴버를 가리킨다. 이후 맴버가 구현되면 고쳐야됨.
     private static final Long NOT_EXIST_COMMENT_ID = 10L;
-    private static String token = null;
-
-    @Autowired
-    JwtTokenProvider jwtTokenProvider;
-
-    @LocalServerPort
-    int port;
 
     private Map<String, Object> params;
 
     @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-        token = jwtTokenProvider.createToken("123456789");
+    public void setUp() {
+        super.setUp();
 
         createArticle("content", EMOTION_ID, SUB_EMOTION_IDS, true);
